refactor datasource_backend_mode.test.ts
